+++ src/app/app.module.ts
@@ -8,10 +8,7 @@ import { DatePipe } from '@angular/common'
 import { AppComponent }   from './app.component';
 import { routing }   from './app.routing';
 
-import { NavComponent }   from './navigation';
-import { FooterComponent}   from './footer'
-import { HomeComponent }   from './home';
-import { NotFoundComponent} from './not-found';
+import { NavComponent, FooterComponent, HomeComponent, NotFoundComponent } from './common';
 import { SensorsComponent }   from './sensors';
 import { RecordsComponent }   from './records';
 import { PatientsComponent }   from './patients';
