+++ src/app/common/footer.component.ts
