+++ src/app/common/index.ts
@@ -1,0 +1,4 @@
+export * from './nav.component';
+export * from './home.component';
+export * from './not-found.component';
+export * from './footer.component';
