--- src/app/home/index.ts
@@ -1,1 +1,0 @@
-export * from './home.component';
