--- src/app/navigation/index.ts
@@ -1,1 +1,0 @@
-export * from './nav.component';
