--- src/app/not-found/index.ts
@@ -1,1 +1,0 @@
-export * from './not-found.component';
